pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poggio-ai"
version = "1.0.0"
description = "PoggioAI/MSc — End-to-end agentic research system by the Poggio Lab at MIT"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Poggio Lab at MIT" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
]

# Runtime dependencies: engine + CLI
dependencies = [
    # Engine (LangGraph pipeline)
    "langgraph>=0.4.0,<1.0",
    "langchain-core>=0.3.0,<1.0",
    "langchain-community>=0.3.0,<1.0",
    "langchain-openai>=0.3.0,<1.0",
    "langchain-anthropic>=0.3.0,<1.0",
    "litellm>=1.60.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
    "openai>=1.0,<3.0",
    "anthropic>=0.30,<2.0",
    "backoff>=2.2,<3.0",
    "requests>=2.31,<3.0",
    "beautifulsoup4>=4.12,<5.0",
    "pymupdf>=1.24,<2.0",
    "pdfminer.six>=20231228",
    "pypdf>=4.0,<7.0",
    "chardet<6",
    "toml",
    # CLI
    "click>=8.1,<9.0",
    "rich>=13.0,<14.0",
    "questionary>=2.0,<3.0",
]

[project.optional-dependencies]
# pip install poggio-ai[observability]
observability = [
    "langsmith>=0.1.0",
    "wandb",
]

# pip install poggio-ai[docs]
docs = [
    "mammoth==1.9.1",
    "markdownify==1.1.0",
    "python-pptx",
    "puremagic",
    "pydub",
    "SpeechRecognition",
    "youtube-transcript-api",
]

# pip install poggio-ai[web]
web = [
    "crawl4ai",
    "playwright",
    "serpapi",
    "pathvalidate",
]

# pip install poggio-ai[experiment]
experiment = [
    "torch",
    "transformers",
    "datasets",
    "huggingface_hub",
    "tokenizers",
    "tqdm",
    "wandb",
]

# pip install poggio-ai[dev]
dev = [
    "pytest",
    "pytest-cov",
]

# pip install poggio-ai[secure]
secure = [
    "keyring>=24.0",
]

# pip install poggio-ai[all] — everything except vllm / gpu extras
all = [
    "poggio-ai[observability,docs,web,experiment,secure]",
]

[project.scripts]
msc = "consortium.cli.main:cli"
consortium = "consortium.runner:main"

[project.urls]
Homepage = "https://github.com/PoggioAI/PoggioAI_MSc"

[tool.setuptools.packages.find]
where = ["."]
include = ["consortium*"]
